Added some code to handle a bot restart when a pug is in progress.
This addresses the problem of losing the teams/maps and knowing the pug was locked. Have put in place a bit of a hack to get around the problem for now. Some further work on the GameServer class where the class can return true/false of if the pug should lock or not. Will need a bit more thought when I've got more time.