Fixed bug where "None" were being written to the signed players list.
This would happen where a match is in progress and the config file gets written to (e.g. because the bot was restart or a cog reloaded), so the players list contains "None" from where they were moved to the team lists.

Not really happy with the solution as when the pug reset it'll unsign an empty list. This is an edge case though, so will leave it for now.